RAG/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-text-splitters
pymupdf
numpy
httpx
openai